--- src/components/MoviesButtonNav.js
@@ -1,12 +1,0 @@
-import React from 'react'
-import CustomButton from './CustomButton'
-
-const MoviesButtonNav = () => {
-  return (
-    <CustomButton>
-      Movies
-    </CustomButton>
-  )
-}
-
-export default MoviesButtonNav
