--- src/components/PeopleButtonNav.js
@@ -1,12 +1,0 @@
-import React from 'react'
-import CustomButton from './CustomButton'
-
-const PeopleButtonNav = () => {
-  return (
-    <CustomButton>
-      People
-    </CustomButton>
-  )
-}
-
-export default PeopleButtonNav
