--- src/components/PlanetsButtonNav.js
@@ -1,12 +1,0 @@
-import React from 'react'
-import CustomButton from './CustomButton'
-
-const PlanetsButtonNav = () => {
-  return (
-    <CustomButton>
-      Planets
-    </CustomButton>
-  )
-}
-
-export default PlanetsButtonNav
